% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson2d.R
\name{cpoisson2d}
\alias{cpoisson2d}
\alias{poisson2d}
\title{Generate poisson disc samples}
\usage{
cpoisson2d(w = 400, h = 300, r = 10, k = 30L, verbosity = 0L)

poisson2d(w = 400, h = 300, r = 10, k = 30L, verbosity = 0L)
}
\arguments{
\item{w, h}{width and height of region}

\item{r}{minimum distance between points}

\item{k}{number of sample points to generate at each iteration. default 30}

\item{verbosity}{Verbosity level. default: 0}
}
\value{
data.frame with x and y coordinates and the 'idx' order in which
        the points were added.
}
\description{
Generate poisson disc samples
}
