% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson3d}
\alias{poisson3d}
\title{Generate Poisson disk samples in 3D}
\usage{
poisson3d(w = 10, h = 10, d = 10, r = 4, k = 30L, verbosity = 0L)
}
\arguments{
\item{w, h, d}{width and height and depth of region}

\item{r}{minimum distance between points}

\item{k}{number of sample points to generate at each iteration. default 30}

\item{verbosity}{Verbosity level. default: 0}
}
\value{
data.frame with x, y and z coordinates. Points are returned in 
    the order in which they were generated.
}
\description{
Generate Poisson disk samples in 3D
}
\examples{
poisson3d(w = 10, h = 10, d = 10, r = 5)
}
