% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_disc.R
\name{poisson_disc}
\alias{poisson_disc}
\title{Generate poisson disc samples}
\usage{
poisson_disc(
  ncols = 20L,
  nrows = 20L,
  cell_size = 10,
  k = 30L,
  xinit = NULL,
  yinit = NULL,
  keep_idx = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ncols, nrows}{number of cells in x and y direction. Resulting point coordinate
ranges will be x = [0, ncols*cell_size] and y = [0, nrows*cell_size]}

\item{cell_size}{length of side of an individual cell}

\item{k}{number of sample points to generate at each iteration. default 30}

\item{xinit, yinit}{coordinates of seed point. If either is NULL, then a random point
will be chosen. Default: NULL}

\item{keep_idx}{return an index indicating the order in which the points were
calcualted. default: FALSE}

\item{verbose}{default: FALSE}
}
\value{
data.frame with x and y coordinates and the 'idx' order in which
        the points were added.
}
\description{
Generate poisson disc samples
}
