% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissoned-package.R
\docType{package}
\name{poissoned-package}
\alias{poissoned}
\alias{poissoned-package}
\title{poissoned: Poisson Disk Sampling in 2D and 3D}
\description{
Poisson disk sampling is a method of generating blue noise sample patterns where all samples are at least a specified distance apart. Poisson samples may be generated in two or three dimensions with this package. The algorithm used is an implementation of Bridson (2007) "Fast Poisson disk sampling in arbitrary dimensions" \doi{10.1145/1278780.1278807}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/poissoned}
  \item Report bugs at \url{https://github.com/coolbutuseless/poissoned/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
