% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_disc.R
\name{poisson_disc}
\alias{poisson_disc}
\title{Generate tileable poisson disc samples}
\usage{
poisson_disc(ncols = 20L, nrows = 20L, cell_size = 10, k = 30L,
  tileable = TRUE, keep_boundary = FALSE, verbose = FALSE)
}
\arguments{
\item{ncols, nrows}{number of cells in x and y direction. Resulting point coordinate
ranges will be x = [0, ncols*cell_size] and y = [0, nrows*cell_size]}

\item{cell_size}{length of side of an individual cell}

\item{k}{number of sample points to generate at each iteration. default 30}

\item{tileable}{enforce boundary conditions to make point set tileable. default: TRUE}

\item{keep_boundary}{keep the boundary condition points used for enforcing
tileability. This is probably only useful for debugging.  default: FALSE}

\item{verbose}{default: FALSE}
}
\value{
data.frame with x and y coordinates
}
\description{
Generate tileable poisson disc samples
}
